requirements.txt
# Core dependencies
anndata>=0.9.0
scanpy>=1.9.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Database dependencies
SQLAlchemy>=1.4.0
alembic>=1.7.0
sqlite-utils>=3.0

# Utility libraries
psutil>=5.8.0
tqdm>=4.62.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Testing
pytest>=6.0.0
pytest-asyncio>=0.18.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Optional dependencies (uncomment as needed)
# scrublet>=0.2.3  # For doublet detection
# umap-learn>=0.5.0  # For dimensionality reduction
# harmony-pytorch>=0.1.0  # For batch correction

# Workflow management
networkx>=3.3
sqlalchemy>=2.0.41
psutil>=5.9.8

# Async/parallel processing
anyio>=4.9.0
aiohttp>=3.9.0
concurrent-log-handler>=0.9.24

# Development/testing
ruff>=0.5.7
black>=24.10.0

# Web & Reporting
fastapi>=0.100.0,<0.112.0  # Example for potential web dashboard, if metrics_collector.py is to be fully implemented
uvicorn[standard]>=0.20.0,<0.30.0 # For serving FastAPI
# mako is a dependency of alembic, usually not needed directly unless templating outside alembic
Mako>=1.2.0,<1.4.0

# Typing
typing-extensions>=4.0.0,<4.14.0 # Often needed by modern libraries

# LLM & AI Tools (Initial versions, can be updated)
langchain>=0.1.0,<0.3.0
dspy-ai>=0.1.0,<0.2.0 # Note: dspy-ai is the package name for DSPy
llm>=0.13,<0.14

# Data Exploration & Utilities
datasette>=0.64,<0.65

# Development & Testing Tools
isort>=5.12.0,<5.14.0 